+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix exception in `req.fresh`/`req.stale` without response headers
+
 4.10.5 / 2014-12-10
 ===================
 
@@ -607,6 +612,21 @@
    - `app.router()` - returns the app Router instance
    - `app.route()` - Proxy to the app's `Router#route()` method to create a new route
    - Router & Route - public API
+
+3.18.6 / 2014-12-12
+===================
+
+  * Fix exception in `req.fresh`/`req.stale` without response headers
+
+3.18.5 / 2014-12-11
+===================
+
+  * deps: connect@2.27.6
+    - deps: compression@~1.2.2
+    - deps: express-session@~1.9.3
+    - deps: http-errors@~1.2.8
+    - deps: serve-index@~1.5.3
+    - deps: type-is@~1.5.4
 
 3.18.4 / 2014-11-23
 ===================